Updated Makefile.

+++ Makefile
@@ -21,7 +21,7 @@ test: $(TARGET)
 	$(FC) $(FFLAGS) $(LDFLAGS) -o test_zlib test/test_zlib.f90 $(TARGET) $(LDLIBS)
 
 clean:
-	rm *.o
-	rm *.mod
-	rm $(TARGET)
-	rm test_zlib
+	if [ `ls -1 *.mod 2>/dev/null | wc -l` -gt 0 ]; then rm *.mod; fi
+	if [ `ls -1 *.o 2>/dev/null | wc -l` -gt 0 ]; then rm *.o; fi
+	if [ -e test_zlib ]; then rm test_zlib; fi
+	if [ -e $(TARGET) ]; then rm $(TARGET); fi
